Added callbacks and improved evaluators.
